typo corrected

+++ app/latticeField_tests/print_tests.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 
 int main() {
-    LatticeField<float> latticeFloat(4, 4, 4, 4, 0.0);
+    LatticeField<double> latticeFloat(4, 4, 4, 4, 0.0);
     std::cout << "latticeFloat(0,1,2,3) : " << latticeFloat(0,1,2,3) << std::endl;
 
     LatticeField<double> latticeDouble(4, 4, 4, 4, 1.0);
